Add test cases for some functions
Add tests for:
- get1stExt
- replaceExt
- hasExt
- LexicographicSorter
- endsWidth
- expandEnv
- getFilename (for IShellItem)

Move some code to separate headers/sources to make testing easier:
- CommonUI.h for some Windows UI
- FS.h for filesystem and path
- COM.h
- StringUtils.h
This doesn't change the functionality in any way.

Update UnitTests.vcxproj to match the settings in notepadPlus.vcxproj.
Add AdditionalIncludeDirectories for all configurations in tests.

Fix x32 warning and treat all warnings as errors.
