+++ PowerEditor/src/MISC/Common/COM.cpp
@@ -1,0 +1,33 @@
+// This file is part of Notepad++ project
+// Copyright (C) 2021 The Notepad++ Contributors.
+
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// at your option any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+#include "COM.h"
+
+generic_string getFilename(IShellItem* psi)
+{
+	generic_string result;
+	if (psi)
+	{
+		PWSTR pszFilePath = nullptr;
+		HRESULT hr = psi->GetDisplayName(SIGDN_FILESYSPATH, &pszFilePath);
+		if (SUCCEEDED(hr) && pszFilePath)
+		{
+			result = pszFilePath;
+			CoTaskMemFree(pszFilePath);
+		}
+	}
+	return result;
+}
