+++ PowerEditor/src/MISC/Common/COM.h
@@ -1,0 +1,38 @@
+// This file is part of Notepad++ project
+// Copyright (C) 2021 The Notepad++ Contributors.
+
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// at your option any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+#pragma once
+
+#include "StringUtils.h"
+#include <comdef.h>		// _com_error
+#include <comip.h>		// _com_ptr_t
+#include <shobjidl.h>
+
+// Utility for getting UID of the interface T.
+// Workaround for MinGW because its implementation of __uuidof is different.
+template<class T>
+struct ComTraits
+{
+	static const GUID uid;
+};
+template<class T>
+const GUID ComTraits<T>::uid = __uuidof(T);
+
+// Smart pointer alias for COM objects that makes reference counting easier.
+template<class T, class InterfaceT = T>
+using com_ptr = _com_ptr_t<_com_IIID<T, &ComTraits<InterfaceT>::uid>>;
+
+generic_string getFilename(IShellItem* psi);
