+++ PowerEditor/src/MISC/Common/Common.cpp
@@ -22,7 +22,6 @@
 #include <locale>
 
 #include "StaticDialog.h"
-#include "CustomFileDialog.h"
 #include "Common.h"
 #include "Utf8.h"
 #include <Parameters.h>
@@ -127,39 +126,6 @@ void writeLog(const TCHAR *logFileName, const char *log2write)
 	fputc('\n', f);
 	fflush(f);
 	fclose(f);
-}
-
-
-generic_string folderBrowser(HWND parent, const generic_string & title, int outputCtrlID, const TCHAR *defaultStr)
-{
-	generic_string folderName;
-	CustomFileDialog dlg(parent);
-	dlg.setTitle(title.c_str());
-
-	// Get an initial directory from the edit control or from argument provided
-	TCHAR directory[MAX_PATH] = {};
-	if (outputCtrlID != 0)
-		::GetDlgItemText(parent, outputCtrlID, directory, _countof(directory));
-	directory[_countof(directory) - 1] = '\0';
-	if (!directory[0] && defaultStr)
-		dlg.setFolder(defaultStr);
-	else if (directory[0])
-		dlg.setFolder(directory);
-
-	folderName = dlg.pickFolder();
-	if (!folderName.empty())
-	{
-		// Send the result back to the edit control
-		if (outputCtrlID != 0)
-			::SetDlgItemText(parent, outputCtrlID, folderName.c_str());
-	}
-	return folderName;
-}
-
-
-generic_string getFolderName(HWND parent, const TCHAR *defaultDir)
-{
-	return folderBrowser(parent, TEXT("Select a folder"), 0, defaultDir);
 }
 
 
@@ -970,98 +936,6 @@ generic_string GetLastErrorAsString(DWORD errorCode)
 	return errorMsg;
 }
 
-HWND CreateToolTip(int toolID, HWND hDlg, HINSTANCE hInst, const PTSTR pszText)
-{
-	if (!toolID || !hDlg || !pszText)
-	{
-		return NULL;
-	}
-
-	// Get the window of the tool.
-	HWND hwndTool = GetDlgItem(hDlg, toolID);
-	if (!hwndTool)
-	{
-		return NULL;
-	}
-
-	// Create the tooltip. g_hInst is the global instance handle.
-	HWND hwndTip = CreateWindowEx(0, TOOLTIPS_CLASS, NULL,
-		WS_POPUP | TTS_ALWAYSTIP | TTS_BALLOON,
-		CW_USEDEFAULT, CW_USEDEFAULT,
-		CW_USEDEFAULT, CW_USEDEFAULT,
-		hDlg, NULL,
-		hInst, NULL);
-
-	if (!hwndTip)
-	{
-		return NULL;
-	}
-
-	NppDarkMode::setDarkTooltips(hwndTip, NppDarkMode::ToolTipsType::tooltip);
-
-	// Associate the tooltip with the tool.
-	TOOLINFO toolInfo = { 0 };
-	toolInfo.cbSize = sizeof(toolInfo);
-	toolInfo.hwnd = hDlg;
-	toolInfo.uFlags = TTF_IDISHWND | TTF_SUBCLASS;
-	toolInfo.uId = (UINT_PTR)hwndTool;
-	toolInfo.lpszText = pszText;
-	if (!SendMessage(hwndTip, TTM_ADDTOOL, 0, (LPARAM)&toolInfo))
-	{
-		DestroyWindow(hwndTip);
-		return NULL;
-	}
-
-	SendMessage(hwndTip, TTM_ACTIVATE, TRUE, 0);
-	SendMessage(hwndTip, TTM_SETMAXTIPWIDTH, 0, 200);
-	// Make tip stay 15 seconds
-	SendMessage(hwndTip, TTM_SETDELAYTIME, TTDT_AUTOPOP, MAKELPARAM((15000), (0)));
-
-	return hwndTip;
-}
-
-HWND CreateToolTipRect(int toolID, HWND hWnd, HINSTANCE hInst, const PTSTR pszText, const RECT rc)
-{
-	if (!toolID || !hWnd || !pszText)
-	{
-		return NULL;
-	}
-
-	// Create the tooltip. g_hInst is the global instance handle.
-	HWND hwndTip = CreateWindowEx(0, TOOLTIPS_CLASS, NULL,
-		WS_POPUP | TTS_ALWAYSTIP | TTS_BALLOON,
-		CW_USEDEFAULT, CW_USEDEFAULT,
-		CW_USEDEFAULT, CW_USEDEFAULT,
-		hWnd, NULL,
-		hInst, NULL);
-
-	if (!hwndTip)
-	{
-		return NULL;
-	}
-
-	// Associate the tooltip with the tool.
-	TOOLINFO toolInfo = { 0 };
-	toolInfo.cbSize = sizeof(toolInfo);
-	toolInfo.hwnd = hWnd;
-	toolInfo.uFlags = TTF_SUBCLASS;
-	toolInfo.uId = toolID;
-	toolInfo.lpszText = pszText;
-	toolInfo.rect = rc;
-	if (!SendMessage(hwndTip, TTM_ADDTOOL, 0, (LPARAM)&toolInfo))
-	{
-		DestroyWindow(hwndTip);
-		return NULL;
-	}
-
-	SendMessage(hwndTip, TTM_ACTIVATE, TRUE, 0);
-	SendMessage(hwndTip, TTM_SETMAXTIPWIDTH, 0, 200);
-	// Make tip stay 15 seconds
-	SendMessage(hwndTip, TTM_SETDELAYTIME, TTDT_AUTOPOP, MAKELPARAM((15000), (0)));
-
-	return hwndTip;
-}
-
 bool isCertificateValidated(const generic_string & fullFilePath, const generic_string & subjectName2check)
 {
 	bool isOK = false;
@@ -1318,4 +1192,26 @@ int nbDigitsFromNbLines(size_t nbLines)
 		}
 	}
 	return nbDigits;
+}
+
+void expandEnv(generic_string& s)
+{
+	TCHAR buffer[MAX_PATH] = { 0 };
+	// This returns the resulting string length or 0 in case of error.
+	DWORD ret = ExpandEnvironmentStrings(s.c_str(), buffer, static_cast<DWORD>(std::size(buffer)));
+	if (ret != 0)
+	{
+		if (ret == static_cast<DWORD>(lstrlen(buffer) + 1))
+		{
+			s = buffer;
+		}
+		else
+		{
+			// Buffer was too small, try with a bigger buffer of the required size.
+			std::vector<TCHAR> buffer2(ret, 0);
+			ret = ExpandEnvironmentStrings(s.c_str(), buffer2.data(), static_cast<DWORD>(buffer2.size()));
+			assert(ret == static_cast<DWORD>(lstrlen(buffer2.data()) + 1));
+			s = buffer2.data();
+		}
+	}
 }
