+++ PowerEditor/visual.net/notepadPlus.vcxproj
@@ -403,6 +403,10 @@ copy ..\src\contextMenu.xml ..\binarm64\contextMenu.xml
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="..\src\DarkMode\DarkMode.cpp" />
+    <ClCompile Include="..\src\MISC\Common\COM.cpp" />
+    <ClCompile Include="..\src\MISC\Common\CommonUI.cpp" />
+    <ClCompile Include="..\src\MISC\Common\FS.cpp" />
+    <ClCompile Include="..\src\MISC\Common\StringUtils.cpp" />
     <ClCompile Include="..\src\MISC\Common\verifySignedfile.cpp" />
     <ClCompile Include="..\src\MISC\md5\md5Dlgs.cpp" />
     <ClCompile Include="..\src\MISC\sha2\sha-256.cpp" />
@@ -675,7 +679,11 @@ copy ..\src\contextMenu.xml ..\binarm64\contextMenu.xml
     <ClInclude Include="..\src\dpiManager.h" />
     <ClInclude Include="..\src\keys.h" />
     <ClInclude Include="..\src\localizationString.h" />
+    <ClInclude Include="..\src\MISC\Common\COM.h" />
+    <ClInclude Include="..\src\MISC\Common\CommonUI.h" />
+    <ClInclude Include="..\src\MISC\Common\FS.h" />
     <ClInclude Include="..\src\MISC\Common\Sorters.h" />
+    <ClInclude Include="..\src\MISC\Common\StringUtils.h" />
     <ClInclude Include="..\src\MISC\Common\verifySignedfile.h" />
     <ClInclude Include="..\src\MISC\md5\md5.h" />
     <ClInclude Include="..\src\MISC\md5\md5Dlgs.h" />
